+++ apis/sgv2-quarkus-common/src/main/java/io/stargate/sgv2/api/common/exception/StatusRuntimeExceptionMapper.java
@@ -19,6 +19,7 @@ package io.stargate.sgv2.api.common.exception;
 
 import io.grpc.Status;
 import io.grpc.StatusRuntimeException;
+import io.quarkus.arc.lookup.LookupIfProperty;
 import io.stargate.sgv2.api.common.exception.model.dto.ApiError;
 import javax.ws.rs.core.Response;
 import org.jboss.resteasy.reactive.RestResponse;
@@ -28,6 +29,10 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /** Simple exception mapper for the {@link StatusRuntimeException}. */
+@LookupIfProperty(
+    name = "stargate.exception-mappers.enabled",
+    stringValue = "true",
+    lookupIfMissing = true)
 public class StatusRuntimeExceptionMapper {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(StatusRuntimeExceptionMapper.class);
@@ -67,12 +72,14 @@ public class StatusRuntimeExceptionMapper {
       case UNIMPLEMENTED:
         return grpcResponseAndLogAsError(
             sc, Response.Status.NOT_IMPLEMENTED, "Unimplemented gRPC operation", msg);
+      case DEADLINE_EXCEEDED:
+        return grpcResponseAndLogAsError(
+            sc, Response.Status.GATEWAY_TIMEOUT, "gRPC service timeout", msg);
 
         // and then all codes we consider as not expected
       case OK:
       case CANCELLED:
       case UNKNOWN:
-      case DEADLINE_EXCEEDED:
       case ALREADY_EXISTS:
       case RESOURCE_EXHAUSTED:
       case ABORTED: