+++ apis/sgv2-quarkus-common/src/main/java/io/stargate/sgv2/api/common/security/HeaderBasedAuthenticationMechanism.java
@@ -17,8 +17,6 @@
 
 package io.stargate.sgv2.api.common.security;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import io.netty.handler.codec.http.HttpResponseStatus;
 import io.quarkus.security.identity.IdentityProviderManager;
 import io.quarkus.security.identity.SecurityIdentity;
@@ -28,13 +26,11 @@ import io.quarkus.vertx.http.runtime.security.HttpAuthenticationMechanism;
 import io.quarkus.vertx.http.runtime.security.HttpCredentialTransport;
 import io.quarkus.vertx.http.runtime.security.HttpSecurityUtils;
 import io.smallrye.mutiny.Uni;
-import io.stargate.sgv2.api.common.exception.model.dto.ApiError;
 import io.vertx.ext.web.RoutingContext;
 import java.util.Collections;
 import java.util.Optional;
 import java.util.Set;
-import javax.ws.rs.core.HttpHeaders;
-import javax.ws.rs.core.MediaType;
+import javax.enterprise.inject.Instance;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -52,12 +48,16 @@ public class HeaderBasedAuthenticationMechanism implements HttpAuthenticationMec
   /** The name of the header to be used for the authentication. */
   private final String headerName;
 
-  /** Object mapper for custom response. */
-  private final ObjectMapper objectMapper;
+  /** Customize the challenge. */
+  private final Instance<io.stargate.sgv2.api.common.security.challenge.ChallengeSender>
+      customChallengeSender;
 
-  public HeaderBasedAuthenticationMechanism(String headerName, ObjectMapper objectMapper) {
+  public HeaderBasedAuthenticationMechanism(
+      String headerName,
+      Instance<io.stargate.sgv2.api.common.security.challenge.ChallengeSender>
+          customChallengeSender) {
     this.headerName = headerName;
-    this.objectMapper = objectMapper;
+    this.customChallengeSender = customChallengeSender;
   }
 
   /** {@inheritDoc} */
@@ -91,43 +91,12 @@ public class HeaderBasedAuthenticationMechanism implements HttpAuthenticationMec
    */
   @Override
   public Uni<Boolean> sendChallenge(RoutingContext context) {
-    // keep the original flow, although I don't need the getChallenge
-    return getChallenge(context)
-        .flatMap(
-            challengeData -> {
-              int status = challengeData.status;
-              context.response().setStatusCode(status);
+    // if we should not customize use default
+    if (!customChallengeSender.isResolvable()) {
+      return HttpAuthenticationMechanism.super.sendChallenge(context);
+    }
 
-              // create the response
-              String message =
-                  "Role unauthorized for operation: Missing token, expecting one in the %s header."
-                      .formatted(headerName);
-              ApiError apiError = new ApiError(message, status);
-              try {
-                // try to serialize
-                String response = objectMapper.writeValueAsString(apiError);
-
-                // set content type
-                context
-                    .response()
-                    .headers()
-                    .set(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON);
-
-                // set content length
-                context
-                    .response()
-                    .headers()
-                    .set(HttpHeaders.CONTENT_LENGTH, String.valueOf(response.getBytes().length));
-
-                // write and map to true
-                return Uni.createFrom()
-                    .completionStage(
-                        context.response().write(response).map(true).toCompletionStage());
-              } catch (JsonProcessingException e) {
-                LOG.error("Unable to serialize API error instance {} to JSON.", apiError, e);
-                return Uni.createFrom().item(true);
-              }
-            });
+    return getChallenge(context).flatMap(c -> customChallengeSender.get().apply(context, c));
   }
 
   /** {@inheritDoc} */