+++ apis/sgv2-restapi/src/test/java/io/stargate/sgv2/it/RestApiV2QIntegrationTestBase.java
@@ -16,7 +16,6 @@ import io.stargate.sgv2.api.common.config.constants.HttpConstants;
 import io.stargate.sgv2.api.common.cql.builder.CollectionIndexingType;
 import io.stargate.sgv2.common.IntegrationTestUtils;
 import io.stargate.sgv2.restapi.service.models.Sgv2ColumnDefinition;
-import io.stargate.sgv2.restapi.service.models.Sgv2GetResponse;
 import io.stargate.sgv2.restapi.service.models.Sgv2IndexAddRequest;
 import io.stargate.sgv2.restapi.service.models.Sgv2RESTResponse;
 import io.stargate.sgv2.restapi.service.models.Sgv2Table;
@@ -31,20 +30,43 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 import org.apache.http.HttpStatus;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.ClassOrderer;
+import org.junit.jupiter.api.MethodOrderer;
+import org.junit.jupiter.api.Order;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestClassOrder;
 import org.junit.jupiter.api.TestInfo;
 import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.api.TestMethodOrder;
 
+/**
+ * Serves as the base class for integration tests that need to create namespace prior to running the
+ * tests and also provides shared convenience methods
+ *
+ * <p>Note that due to the way how RestAssured is configured in Quarkus tests, we are doing the
+ * initialization as first tests to be run. The {@link BeforeAll} annotation can not be used, as
+ * rest assured is not configured yet. See https://github.com/quarkusio/quarkus/issues/7690 for more
+ * info.
+ */
 @TestClassOrder(ClassOrderer.DisplayName.class)
+@TestMethodOrder(MethodOrderer.OrderAnnotation.class)
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 public abstract class RestApiV2QIntegrationTestBase {
+  public enum KeyspaceCreation {
+    PER_METHOD,
+    PER_CLASS,
+    NONE
+  }
+
   protected static final ObjectMapper objectMapper = JsonMapper.builder().build();
 
   protected static final TypeReference LIST_OF_MAPS_TYPE =
       new TypeReference<List<Map<String, Object>>>() {};
+
+  private final KeyspaceCreation keyspaceCreation;
 
   private final String testKeyspacePrefix;
 
@@ -60,9 +82,11 @@ public abstract class RestApiV2QIntegrationTestBase {
   /////////////////////////////////////////////////////////////////////////
    */
 
-  protected RestApiV2QIntegrationTestBase(String keyspacePrefix, String tablePrefix) {
+  protected RestApiV2QIntegrationTestBase(
+      String keyspacePrefix, String tablePrefix, KeyspaceCreation keyspaceCreation) {
     this.testKeyspacePrefix = keyspacePrefix;
     this.testTablePrefix = tablePrefix;
+    this.keyspaceCreation = keyspaceCreation;
   }
 
   @BeforeAll
@@ -84,12 +108,40 @@ public abstract class RestApiV2QIntegrationTestBase {
       // need to truncate testName, NOT timestamp, so:
       testName = testName.substring(0, testName.length() - (len - 48));
     }
-    testKeyspaceName = testKeyspacePrefix + testName + timestamp;
     testTableName = testTablePrefix + testName + timestamp;
 
-    // Create keyspace automatically (same as before)
-    createKeyspace(testKeyspaceName);
-    // But not table (won't have definition anyway)
+    // May need automatic keyspace creation on per-class or per-method  basis
+    if (keyspaceCreation == KeyspaceCreation.PER_METHOD) {
+      testKeyspaceName = testKeyspacePrefix + testName + timestamp;
+      createKeyspace(testKeyspaceName);
+    }
+  }
+
+  // Per-class initialization needs to run "as a test method" before real tests
+  @Test
+  @Order(Integer.MIN_VALUE)
+  public void initPerClass() {
+    if (keyspaceCreation == KeyspaceCreation.PER_CLASS) {
+      // If shared for test methods create first time it is actually needed
+      testKeyspaceName = testKeyspacePrefix + "shared";
+      createKeyspace(testKeyspaceName);
+    }
+  }
+
+  @AfterEach
+  public void cleanupPerTest() {
+    if (keyspaceCreation == KeyspaceCreation.PER_METHOD) {
+      deleteKeyspace(testKeyspaceName);
+    }
+  }
+
+  // Per-class cleanup needs to run "as a test method" after all real tests
+  @Test
+  @Order(Integer.MAX_VALUE)
+  public void cleanupPerClass() {
+    if (keyspaceCreation == KeyspaceCreation.PER_CLASS) {
+      deleteKeyspace(testKeyspaceName);
+    }
   }
 
   /*
@@ -123,6 +175,10 @@ public abstract class RestApiV2QIntegrationTestBase {
   // Endpoint construction
   /////////////////////////////////////////////////////////////////////////
    */
+
+  protected String endpointPathForCQL() {
+    return "/v2/cql";
+  }
 
   protected String endpointPathForAllKeyspaces() {
     return "/v2/schemas/keyspaces";
@@ -201,7 +257,7 @@ public abstract class RestApiV2QIntegrationTestBase {
         objectMapper.getTypeFactory().constructParametricType(Sgv2RESTResponse.class, wrappedType);
     try {
       Sgv2RESTResponse<T> wrapped = objectMapper.readValue(body, wrapperType);
-      return wrapped.getData();
+      return wrapped.data();
     } catch (IOException e) {
       throw new RuntimeException(e);
     }
@@ -245,10 +301,10 @@ public abstract class RestApiV2QIntegrationTestBase {
   /////////////////////////////////////////////////////////////////////////
    */
 
-  protected static class ListOfMapsGetResponseWrapper
-      extends Sgv2GetResponse<List<Map<String, Object>>> {
+  protected record ListOfMapsGetResponseWrapper(
+      int count, String pageState, List<Map<String, Object>> data) {
     public ListOfMapsGetResponseWrapper() {
-      super(-1, null, null);
+      this(-1, null, null);
     }
   }
 
@@ -258,7 +314,7 @@ public abstract class RestApiV2QIntegrationTestBase {
 
   /*
   /////////////////////////////////////////////////////////////////////////
-  // Helper methods for Keyspace creation
+  // Helper methods for Keyspace creation/deletion
   /////////////////////////////////////////////////////////////////////////
    */
 
@@ -281,6 +337,14 @@ public abstract class RestApiV2QIntegrationTestBase {
         .statusCode(HttpStatus.SC_CREATED);
   }
 
+  protected void deleteKeyspace(String keyspaceName) {
+    givenWithAuth()
+        .when()
+        .delete(endpointPathForAllKeyspaces() + "/{keyspace-id}", keyspaceName)
+        .then()
+        .statusCode(HttpStatus.SC_NO_CONTENT);
+  }
+
   /*
   /////////////////////////////////////////////////////////////////////////
   // Helper methods for Table Creation, Access
@@ -288,33 +352,30 @@ public abstract class RestApiV2QIntegrationTestBase {
    */
 
   protected NameResponse createSimpleTestTable(String keyspaceName, String tableName) {
-    final Sgv2TableAddRequest tableAdd = new Sgv2TableAddRequest(tableName);
-    tableAdd.setColumnDefinitions(
+    List<Sgv2ColumnDefinition> columnDefs =
         Arrays.asList(
             new Sgv2ColumnDefinition("id", "uuid", false),
             new Sgv2ColumnDefinition("lastName", "text", false),
             new Sgv2ColumnDefinition("firstName", "text", false),
-            new Sgv2ColumnDefinition("age", "int", false)));
+            new Sgv2ColumnDefinition("age", "int", false));
 
-    Sgv2Table.PrimaryKey primaryKey = new Sgv2Table.PrimaryKey();
-    primaryKey.setPartitionKey(Arrays.asList("id"));
-    tableAdd.setPrimaryKey(primaryKey);
-
+    Sgv2Table.PrimaryKey primaryKey = new Sgv2Table.PrimaryKey(Arrays.asList("id"));
+    final Sgv2TableAddRequest tableAdd =
+        new Sgv2TableAddRequest(tableName, primaryKey, columnDefs, false, null);
     return createTable(keyspaceName, tableAdd);
   }
 
   protected NameResponse createComplexTestTable(String keyspaceName, String tableName) {
-    final Sgv2TableAddRequest tableAdd = new Sgv2TableAddRequest(tableName);
-    tableAdd.setColumnDefinitions(
+    List<Sgv2ColumnDefinition> columnDefs =
         Arrays.asList(
             new Sgv2ColumnDefinition("pk0", "uuid", false),
             new Sgv2ColumnDefinition("col1", "frozen<map<date, text>>", false),
             new Sgv2ColumnDefinition("col2", "frozen<set<boolean>>", false),
-            new Sgv2ColumnDefinition("col3", "tuple<duration, inet>", false)));
+            new Sgv2ColumnDefinition("col3", "tuple<duration, inet>", false));
 
-    Sgv2Table.PrimaryKey primaryKey = new Sgv2Table.PrimaryKey();
-    primaryKey.setPartitionKey(Arrays.asList("pk0"));
-    tableAdd.setPrimaryKey(primaryKey);
+    Sgv2Table.PrimaryKey primaryKey = new Sgv2Table.PrimaryKey(Arrays.asList("pk0"));
+    final Sgv2TableAddRequest tableAdd =
+        new Sgv2TableAddRequest(tableName, primaryKey, columnDefs, false, null);
 
     return createTable(keyspaceName, tableAdd);
   }
@@ -325,21 +386,14 @@ public abstract class RestApiV2QIntegrationTestBase {
       List<String> columns,
       List<String> partitionKey,
       List<String> clusteringKey) {
-    Sgv2TableAddRequest tableAdd = new Sgv2TableAddRequest(tableName);
-
-    List<Sgv2ColumnDefinition> columnDefinitions =
+    List<Sgv2ColumnDefinition> columnDefs =
         columns.stream()
             .map(x -> x.split(" "))
             .map(y -> new Sgv2ColumnDefinition(y[0], y[1], false))
             .collect(Collectors.toList());
-    tableAdd.setColumnDefinitions(columnDefinitions);
-
-    Sgv2Table.PrimaryKey primaryKey = new Sgv2Table.PrimaryKey();
-    primaryKey.setPartitionKey(partitionKey);
-    if (clusteringKey != null) {
-      primaryKey.setClusteringKey(clusteringKey);
-    }
-    tableAdd.setPrimaryKey(primaryKey);
+    Sgv2Table.PrimaryKey primaryKey = new Sgv2Table.PrimaryKey(partitionKey, clusteringKey);
+    final Sgv2TableAddRequest tableAdd =
+        new Sgv2TableAddRequest(tableName, primaryKey, columnDefs, false, null);
     return createTable(keyspaceName, tableAdd);
   }
 
@@ -542,18 +596,17 @@ public abstract class RestApiV2QIntegrationTestBase {
 
   /** @return Partition key of the first row */
   protected Integer setupClusteringTestCase(String keyspaceName, String tableName) {
-    final Sgv2TableAddRequest tableAdd = new Sgv2TableAddRequest(tableName);
-    tableAdd.setColumnDefinitions(
+    List<Sgv2ColumnDefinition> columnDefs =
         Arrays.asList(
             new Sgv2ColumnDefinition("id", "int", false),
             new Sgv2ColumnDefinition("lastName", "text", false),
             new Sgv2ColumnDefinition("firstName", "text", false),
             new Sgv2ColumnDefinition("age", "int", true),
-            new Sgv2ColumnDefinition("expense_id", "int", false)));
-    Sgv2Table.PrimaryKey primaryKey = new Sgv2Table.PrimaryKey();
-    primaryKey.setPartitionKey(Arrays.asList("id"));
-    primaryKey.setClusteringKey(Arrays.asList("expense_id"));
-    tableAdd.setPrimaryKey(primaryKey);
+            new Sgv2ColumnDefinition("expense_id", "int", false));
+    Sgv2Table.PrimaryKey primaryKey =
+        new Sgv2Table.PrimaryKey(Arrays.asList("id"), Arrays.asList("expense_id"));
+    final Sgv2TableAddRequest tableAdd =
+        new Sgv2TableAddRequest(tableName, primaryKey, columnDefs, false, null);
     createTable(keyspaceName, tableAdd);
 
     final Integer firstRowId = 1;
@@ -586,19 +639,18 @@ public abstract class RestApiV2QIntegrationTestBase {
   }
 
   protected void setupMixedClusteringTestCase(String keyspaceName, String tableName) {
-    final Sgv2TableAddRequest tableAdd = new Sgv2TableAddRequest(tableName);
-    tableAdd.setColumnDefinitions(
+    List<Sgv2ColumnDefinition> columnDefs =
         Arrays.asList(
             new Sgv2ColumnDefinition("pk0", "int", false),
             new Sgv2ColumnDefinition("pk1", "text", false),
             new Sgv2ColumnDefinition("pk2", "int", false),
             new Sgv2ColumnDefinition("ck0", "int", false),
             new Sgv2ColumnDefinition("ck1", "text", false),
-            new Sgv2ColumnDefinition("v", "int", false)));
-    Sgv2Table.PrimaryKey primaryKey = new Sgv2Table.PrimaryKey();
-    primaryKey.setPartitionKey(Arrays.asList("pk0", "pk1", "pk2"));
-    primaryKey.setClusteringKey(Arrays.asList("ck0", "ck1"));
-    tableAdd.setPrimaryKey(primaryKey);
+            new Sgv2ColumnDefinition("v", "int", false));
+    Sgv2Table.PrimaryKey primaryKey =
+        new Sgv2Table.PrimaryKey(Arrays.asList("pk0", "pk1", "pk2"), Arrays.asList("ck0", "ck1"));
+    final Sgv2TableAddRequest tableAdd =
+        new Sgv2TableAddRequest(tableName, primaryKey, columnDefs, false, null);
     createTable(keyspaceName, tableAdd);
 
     Map<String, Object> row = new HashMap<>();